// src/types/enums.ts
export enum Role {
  ADMIN = "ADMIN",
  MESTRE = "MESTRE",
  JOGADOR = "JOGADOR"
}

export enum GameStyle {
  VAMPIRE = "VAMPIRE",
  WEREWOLF = "WEREWOLF",
  MAGE = "MAGE",
  HUNTER = "HUNTER"
}

export enum PowerType {
  DISCIPLINE = "DISCIPLINE",
  GIFT = "GIFT",
  SPHERE = "SPHERE",
  EDGE = "EDGE"
}

export enum MeritFlawType {
  MERIT = "MERIT",
  FLAW = "FLAW"
}